Divided the queries.js file (contains the implementation of the endpoints) into separate files
