+++ endpoints/song.js
@@ -1,9 +1,9 @@
 var express = require('express')
 var router = express.Router()
 
-var { songQueries } = require('../queries')
+var queries = require('../endpointImpl/songQueries')
 
-router.get('/:text', songQueries.songSearchText);
+router.get('/:text', queries.searchText);
 
 // Future endpoints:
 //    /popularity             GET      getPopular
